+++ src/main/java/com/example/demo/SaleRepo.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface SaleRepo {
 
-    ArrayList<Sale> saveSale(String product, Integer quantity, String personName, Float discount, Float mySortPrice);
+    ArrayList<Sale> saveSale(String product, Float quantity, String personName, Float discount, Float mySortPrice);
 
     List<Sale> loadAllSales();
